# api/openapi.yaml
openapi: 3.0.3
info:
  title: PipeRouter Admin API
  version: 0.1.0
  description: |
    REST/JSON management API for PipeRouter, a lightweight HTTP distribution
    proxy. Served under `/api/v1` by the admin server (default
    `127.0.0.1:9090`).

    General rules:
    - All request and response bodies are JSON. Request bodies are limited
      to 1 MiB.
    - No CORS headers are ever emitted. On POST/PUT/DELETE requests that
      carry an `Origin` header, its host must equal the request `Host`,
      otherwise the API responds `403 {"error":"origin_not_allowed"}`.
    - Errors use the shape `{"error":"<code>","detail":"...","issues":[...]}`
      (`issues` only accompanies `validation_failed`). Error codes:
      `validation_failed`, `revision_conflict`, `not_found`,
      `builtin_transport`, `transport_in_use`, `origin_not_allowed`,
      `invalid_request`, `method_not_allowed`, `internal_error`.
    - Configuration mutations accept an optional `revision`; when present it
      must match the active configuration revision or the API responds
      `409 {"error":"revision_conflict"}`. An absent or empty revision skips
      the check.
    - Durations in the configuration are Go duration strings such as
      `"10s"` or `"1m30s"`.
servers:
  - url: /
...
paths:
...
  /api/v1/diagnostics/request:
    post:
      tags: [diagnostics]
      operationId: testRequest
      summary: Probe an inbound request path
      description: |
        Simulates a client request to the data plane: longest-prefix route
        matching on the given path, then rewrite → transport → target, with a
        15-second overall timeout, no redirect following, and at most 64 KiB
        of response body read and discarded. Use this to answer "what would
        happen if a client hit the proxy with this path?" — not as a
        per-route or per-transport health check (those live on the Routes
        and Transports pages). Upstream 401/403/404 still report `ok: true`
        (the link works). Bodies and authentication headers are never stored
        or logged. A path that matches no enabled route returns HTTP 200 with
        `ok: false` and `error_stage: resolve`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestTestRequest'
      responses:
        '200':
          description: Test result (also returned when the probe itself failed; see `ok`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiagnosticsResult'
        '400':
          $ref: '#/components/responses/BadRequestInvalid'
        '403':
          $ref: '#/components/responses/OriginNotAllowed'
  /api/v1/diagnostics/route:
    post:
      tags: [diagnostics]
      operationId: testRoute
      summary: Test a named route end-to-end
      description: |
        Executes the real pipeline (rewrite → transport → target) for the
        named route with a 15-second overall timeout, no redirect following,
        and at most 64 KiB of response body read and discarded. Used by the
        Routes page "Test" action. Upstream 401/403/404 still report
        `ok: true` (the link works). Bodies and authentication headers are
        never stored or logged.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteTestRequest'
      responses:
        '200':
          description: Test result (also returned when the probe itself failed; see `ok`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiagnosticsResult'
        '400':
          $ref: '#/components/responses/BadRequestInvalid'
        '403':
          $ref: '#/components/responses/OriginNotAllowed'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/diagnostics/transport:
    post:
      tags: [diagnostics]
      operationId: testTransport
      summary: Test a transport against a URL
      description: |
        Sends a probe request to the given URL through the named transport
        with a 15-second overall timeout, no redirect following, and at most
        64 KiB of response body read and discarded. The URL must be absolute
        http/https, otherwise `400 invalid_request`. The test URL and
        response body are never stored or logged.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransportTestRequest'
      responses:
        '200':
          description: Test result (also returned when the probe itself failed; see `ok`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiagnosticsResult'
        '400':
          $ref: '#/components/responses/BadRequestInvalid'
        '403':
          $ref: '#/components/responses/OriginNotAllowed'
        '404':
          $ref: '#/components/responses/NotFound'
components:
...
  responses:
...
    BadRequestInvalid:
      description: '`invalid_request` — malformed JSON or an unacceptable field value.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    OriginNotAllowed:
      description: >
        `origin_not_allowed` — the request carried an `Origin` header whose
        host does not match the request `Host`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: '`not_found` — no object with the given name exists.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
...
    RequestTestRequest:
      type: object
      description: >
        Inbound-request diagnostics: path is matched against the live route
        table exactly like the data plane.
      properties:
        path:
          type: string
          description: >
            Full request path as a client would send (must start with `/`;
            empty is treated as `/`).
          example: /openai/models
        method:
          type: string
          description: HTTP method for the probe.
          default: GET
    RouteTestRequest:
      type: object
      description: Named-route diagnostics request (Routes page Test).
      required: [route]
      properties:
        route:
          type: string
          description: Name of the route to test.
          example: openai
        path:
          type: string
          description: Extra path appended below the route prefix before rewriting.
          example: /models
        method:
          type: string
          description: HTTP method for the probe.
          default: GET
    TransportTestRequest:
      type: object
      description: Transport diagnostics request.
      required: [transport, url]
      properties:
        transport:
          type: string
          description: Name of the transport to test ("direct" is allowed).
          example: jp-proxy
        url:
          type: string
          description: Absolute http/https URL to probe.
          example: https://api.openai.com/v1/models
        method:
          type: string
          description: HTTP method for the probe.
          default: GET
    DiagnosticsResult:
      type: object
      description: >
        Outcome of a diagnostics probe. Upstream 401/403/404 responses still
        report `ok: true` — the network link works.
      required:
        - ok
        - route
        - target_url
        - transport
        - status
        - error_stage
        - error
        - header_duration_ms
        - total_duration_ms
      properties:
        ok:
          type: boolean
        route:
          type: string
          description: >
            Matched or named route for the probe; empty when the probe did
            not resolve a route (transport-only tests, or no match).
          example: openai
        target_url:
          type: string
          description: Final rewritten target URL that was probed.
          example: https://api.openai.com/v1/models
        transport:
          type: string
          description: Transport used for the probe.
        status:
          type: integer
          description: Upstream HTTP status; 0 if no HTTP response was received.
        error_stage:
          type: string
          enum: ['', resolve, connect, tls, response]
          description: Stage at which the probe failed; empty on success.
        error:
          type: string
          description: Sanitized error message; empty on success.
        header_duration_ms:
          type: number
          format: double
          description: Time until upstream response headers were received.
        total_duration_ms:
          type: number
          format: double
    ApiError:
      type: object
      description: Standard error envelope.
      required: [error]
      properties:
        error:
          type: string
          description: Machine-readable error code.
          enum:
            - validation_failed
            - revision_conflict
            - not_found
            - builtin_transport
            - transport_in_use
            - origin_not_allowed
            - invalid_request
            - method_not_allowed
            - internal_error
        detail:
          type: string
          description: Optional human-readable detail (never contains credentials or proxy URLs).
        issues:
          type: array
          description: Present only for `validation_failed` — one entry per validation problem.
          items:
            type: string